requirements.txt
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.4.4
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
numpy==1.19.5
pandas==1.1.5
PyMySQL==1.0.2
SQLAlchemy==1.3.23
tqdm==4.56.0